% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ref_seq_for_matches.R
\name{check_ref_seq_for_matches}
\alias{check_ref_seq_for_matches}
\title{Check a vector of sequences for equality on each position}
\usage{
check_ref_seq_for_matches(
  seq_set,
  ref_seq_name,
  pos_col = "position",
  seq_col = "seq",
  name_col = "seq.name"
)
}
\arguments{
\item{seq_set}{XStringSet or character vector of sequences}

\item{ref_seq_name}{name of reference sequence which will not be altered}

\item{pos_col}{name of position column}

\item{seq_col}{name of sequence column}

\item{name_col}{name of column which holds sequence names}
}
\value{
data frame
}
\description{
A vector of sequences is compared to a reference sequence. In case
of equality all sequences except for ref_seq are changed to 'match' or in case
of inequality to 'mismatch'.
}
