% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_duplicate_sequences.R
\name{collapse_duplicate_sequences}
\alias{collapse_duplicate_sequences}
\title{Have a set of sequences checked for duplicates}
\usage{
collapse_duplicate_sequences(seq_set)
}
\arguments{
\item{seq_set}{XStringSet or character vector of sequences}
}
\value{
seq_set with unique sequences only as character vector
}
\description{
Duplicate sequences in the set are collapse to one sequences and named
as 'unique seq i'.
}
