% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_gtf_and_genome_for_cellranger.R
\name{combine_gtf_and_genome_for_cellranger}
\alias{combine_gtf_and_genome_for_cellranger}
\title{Make a custom/hybrid reference for read alignment}
\usage{
combine_gtf_and_genome_for_cellranger(
  genome_files,
  gtf_files,
  fasta_name_two_parts = T,
  save_path = dirname(genome_files[1]),
  save_names = c("genome1.fa", "genes1.gtf"),
  overwrite = F,
  gtf_header =
    c("##description: made with combine_gtf_and_genome_for_cellranger function from https://github.com/Close-your-eyes/igsc",
    paste0("##sources: ", paste(basename(gtf_files), collapse = ", ")), "##creator: CMS",
    "##conctact: vonskopnik_aet_pm.me", "##format: gtf", paste0("##date: ", Sys.Date())),
  verbose = T
)
}
\arguments{
\item{genome_files}{vector of paths to genome fasta files (.fa, .fna, .fasta)}

\item{gtf_files}{vector of paths to matching gtf files (e.g. genes.gtf)}

\item{fasta_name_two_parts}{make names in genome file in 10X Genomics style:
chr1 1 or chrM MT instead of chr1 and chrM only}

\item{save_path}{where to save}

\item{save_names}{save names}

\item{overwrite}{overwrite existing target files?}

\item{gtf_header}{header for gtf}

\item{verbose}{print messages?}
}
\description{
Genome fasta files and gtf files are concatenated respectively and some
checks are run. Pairs of genome_files and gtf_files expected. E.g. seqnames
are tried to be harmonized by an algorithm.
}
\examples{
\dontrun{
# concat files on server:
# remove header from second gtf, then
# cat genes_hg38_renamed.gtf genes_mm10_renamed.gtf > genes.gtf
# cat genome_hg38_renamed.fa genome_mm10_renamed.fa > genome.fa

# hg38 and 5 herpes virus genomes
# consider adding info to gtf header:
##NOTE: CDS features converted to exon for viral genomes; original exons removed
##NOTE: only exons retained from hg38
root <- "/Volumes/CMS_SSD_2TB/reference_genomes/"
viral_dirs <- list.files(paste0(root, "GRCh38_2020_A_HHV1to5_appended/viral_ref_genomes"), pattern = "gtf$", recursive = T, full.names = T)
prefix <- c("", paste0("HHV_", basename(dirname(viral_dirs)), "_"))
combine_gtf_and_genome_for_cellranger(genome_files = c(paste0(root, "refdata-gex-GRCh38-2024-A/fasta/genome.fa.gz"),
                                                       list.files(paste0(root, "GRCh38_2020_A_HHV1to5_appended/viral_ref_genomes"),
                                                                  pattern = "fna$", recursive = T, full.names = T)),
                                      gtf_files = c(paste0(root, "refdata-gex-GRCh38-2024-A/genes/genes.gtf.gz"),
                                                    viral_dirs),
                                      gtf_gene_name_prefix = prefix,
                                      save_path = paste0(root, "GRCh38_2020_A_HHV1to5_appended"),
                                      save_names = c("genome4.fa", "genes4.gtf"),
                                      gtf_features_to_exon = c(list(c("")), rep(list(c("CDS")), 5)),
                                      gtf_rm_exon = c(F, rep(T, 5)))
# then run cellranger mkref
cellranger mkref \
--genome="${GENOME_NAME}" \
--fasta="${FASTA}" \
--genes="${GTF}" \
--nthreads="${SLURM_CPUS_PER_TASK}" \
--memgb=60 \
--output-dir="${OUTDIR}"
}
}
