% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_nonref_r.R
\name{compare_nonref_r}
\alias{compare_nonref_r}
\title{Title}
\usage{
compare_nonref_r(
  df,
  ref,
  pos_col = "position",
  seq_col = "seq",
  name_col = "seq.name",
  match_symbol = ".",
  mismatch_symbol = "x",
  keep_gaps = T,
  nonref_mismatch_as = "base"
)
}
\description{
Title
}
