% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_seq_df_wide.R
\name{compare_seq_df_wide}
\alias{compare_seq_df_wide}
\title{Compare sequences in columns of a data frame}
\usage{
compare_seq_df_wide(
  df,
  ref = NULL,
  non_ref = NULL,
  pos_col = "position",
  seq_col = "seq",
  name_col = "seq.name",
  match_symbol = ".",
  mismatch_symbol = "x",
  change_nonref = F,
  nonref_mismatch_as = c("base", "mismatch_symbol"),
  change_ref = F,
  ref_mismatch_as = c("base", "mismatch_symbol"),
  insertion_as = "base",
  keep_match_gaps = T,
  return_as_long = F,
  rm_pure_NA_non_ref = F
)
}
\arguments{
\item{df}{data frame}

\item{ref}{reference or subject to compare non-ref or patterns to
(col name); can be guessed if NULL}

\item{non_ref}{non-ref or pattern col names; can be guessed if NULL}

\item{pos_col}{column name of position column}

\item{seq_col}{sequence column name when return_as_long}

\item{name_col}{sequence name column name when return_as_long}

\item{match_symbol}{symbol for matching positions}

\item{mismatch_symbol}{symbol for mismatching positions}

\item{change_nonref}{alter non-refs to match, mismatch or insertion}

\item{nonref_mismatch_as}{what should mismatches become in non-ref}

\item{change_ref}{alter ref to match, mismatch or insertion}

\item{ref_mismatch_as}{what should mismatches become in ref}

\item{insertion_as}{what should insertion become}

\item{keep_match_gaps}{keep matching gaps as gaps (T)? of match (F)?}

\item{return_as_long}{return as long data frame?}

\item{rm_pure_NA_non_ref}{if all non-refs are NA at one position, remove this
position? intended to make subsequent plot tighter.}
}
\value{
data frame
}
\description{
Prepare for plotting with algmnt_plot.
}
