% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_exons.R
\name{concat_exons}
\alias{concat_exons}
\title{Pull and concatenate exon ranges into one sequence}
\usage{
concat_exons(gtf_df, refseq)
}
\arguments{
\item{gtf_df}{subsetted gtf data frame}

\item{refseq}{reference sequence to pull from}
}
\description{
Based on gtf_df which should only contain lines of exon definitions for
one gene, the respective sequences are pulled from refseq and concatenated
into the coding sequence (CDS).
}
