% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consecutive_disambiguate_seq.R
\name{consecutive_distinct_seq}
\alias{consecutive_distinct_seq}
\title{Get the longest consecutive stretch of distinct sequence}
\usage{
consecutive_distinct_seq(seq, seq_type = NULL)
}
\arguments{
\item{seq}{Sequence as character, DNAStringSet, DNAString, RNAStringSet, RNAString,
AAStringSet or AAString}

\item{seq_type}{set sequence type to AA or NT if necessary; if NULL
it is attempted to guess the type}
}
\value{
a character sequence
}
\description{
Use case: Have a consensus sequence calculated from a set of DNA or AA sequences. Then
get the longest consecutive stretch of distinct sequence from the consensus.
}
