% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fasta_seq_bounds.R
\name{get_fasta_seq_bounds}
\alias{get_fasta_seq_bounds}
\title{From a genome fasta file, get chromosome sequence boundaries}
\usage{
get_fasta_seq_bounds(file_path)
}
\arguments{
\item{file_path}{path to fasta file}
}
\description{
From a genome fasta file, get chromosome sequence boundaries
}
