% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_exons.R
\name{get_seq_from_refseq}
\alias{get_seq_from_refseq}
\title{Title}
\usage{
get_seq_from_refseq(
  refseq,
  gtf_df,
  group_by = NULL,
  names = NULL,
  what = c("single", "whole")
)
}
\arguments{
\item{what}{}
}
\description{
Title
}
