% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seqs_from_feature_df.R
\name{get_seqs_from_feature_df}
\alias{get_seqs_from_feature_df}
\title{Get sequences based on data generated with webscrape_ncbi}
\usage{
get_seqs_from_feature_df(
  feature_df,
  origin,
  concat = T,
  make_revcomp = F,
  min_to_max = F,
  return = c("sequences", "df_wide", "df_long"),
  matches_to_origin_and_feature = list(c(F, T), c(F, F)),
  order_features = F,
  compare_seq_df_long_args = list()
)
}
\arguments{
\item{feature_df}{data frame of features as returned by webscrape_ncbi}

\item{origin}{reference sequence for features}

\item{concat}{concatenate sequences which have gaps in origin}

\item{make_revcomp}{return the reverse complement of sequences when complement column is TRUE,
for alignment to origin do not have revcomp return,
for writing features into a ref genome do have the it return}

\item{min_to_max}{return whole sequences from start to end including gaps (TRUE), or remove gaps (FALSE)}

\item{return}{which formats to return; restriction to what is needed saves memory and increases speed}

\item{matches_to_origin_and_feature}{mark matches and mismatches in origin and/or feature sequences}

\item{order_features}{order features by min position}

\item{compare_seq_df_long_args}{args to igsc::compare_seq_df_long}
}
\description{
Get sequences based on data generated with webscrape_ncbi
}
