% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sra_ena_aws_links_from_gse.R
\name{get_sra_ena_aws_links_from_gse}
\alias{get_sra_ena_aws_links_from_gse}
\title{Get links to fastq files from GSE accession}
\usage{
get_sra_ena_aws_links_from_gse(
  gse_accession,
  save_file = paste0(gse_accession, ".json"),
  save_dir = tempdir(),
  python_to_PATH = "~/Library/Python/3.13/bin"
)
}
\arguments{
\item{gse_accession}{GEO GSE accession}

\item{save_file}{json file name}

\item{save_dir}{directory for save_file}

\item{python_to_PATH}{python dir on disk to add to PATH}
}
\value{
list
}
\description{
Uses ffq (https://github.com/pachterlab/ffq) and processes the json file
returned.
}
\examples{
\dontrun{
get_sra_ena_aws_links_from_gse(gse_accession = "GSE151302")
}
}
