% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_substitution_matrices.R
\name{get_substitution_matrices}
\alias{get_substitution_matrices}
\title{Load amino acid substitution matrices}
\usage{
get_substitution_matrices()
}
\value{
Called for its side effect of loading matrices.
}
\description{
Loads BLOSUM and PAM matrices from the \code{pwalign} package into the global
environment, then reports which matrices are available.
}
\details{
Existing objects in the global environment with the same names
will be replaced.
}
\examples{
\dontrun{
get_substitution_matrices()
BLOSUM62["A", "G"]
}
}
