% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{igsc-package}
\alias{igsc}
\alias{igsc-package}
\title{tcrdistR}
\description{
Calculate T-cell receptor distances using the Python tcrdist3
package.
}
\author{
\strong{Maintainer}: Christopher Mark Skopnik \email{vonskopnik@pm.me} (\href{https://orcid.org/0000-0002-6089-4769}{ORCID})

Authors:
\itemize{
  \item Christopher Mark Skopnik \email{vonskopnik@pm.me} (\href{https://orcid.org/0000-0002-6089-4769}{ORCID})
}

}
\keyword{internal}
