% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gtf.R
\name{make_kv_attr_col}
\alias{make_kv_attr_col}
\title{Make kv pairs in attribute col for saving proper gtf df}
\usage{
make_kv_attr_col(gtf_df, keep_index_col = F, verbose = T)
}
\arguments{
\item{gtf_df}{subsetted gtf df with processed attribute column}

\item{keep_index_col}{keep index column away from attributes}

\item{verbose}{print messages?}
}
\value{
data frame
}
\description{
From a gtf df with processed attr col, make a attribute column with
key-value pairs
}
\examples{
\dontrun{
gtf <- make_kv_attr_col(gtf)
write_gtf(gtf)
}
}
