% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_short_refseq_from_gtf.R
\name{make_short_refseq_from_gtf}
\alias{make_short_refseq_from_gtf}
\title{Make shorter refseqs from gtf subset and original full refseqs}
\usage{
make_short_refseq_from_gtf(gtf_df, refseqs, overhang_whole = c(0, 0))
}
\arguments{
\item{gtf_df}{subsetted gtf data frame}

\item{refseqs}{named ref sequence character vector; names must be in
gtf_df$seqname.}

\item{overhang_whole}{how much overhang in refseq}
}
\description{
Make shorter refseqs from gtf subset and original full refseqs
}
\examples{
\dontrun{
out <- make_short_refseq_from_gtf(hugtf, genome)
}
}
