% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiplePairwiseAlignmentsToOneSubject.R
\name{MultiplePairwiseAlignmentsToOneSubject}
\alias{MultiplePairwiseAlignmentsToOneSubject}
\title{Align multiple DNA pattern sequences to one subject}
\usage{
MultiplePairwiseAlignmentsToOneSubject(
  subject,
  patterns,
  type = "global-local",
  perfect.matches.only = F,
  order.patterns = F,
  attach.nt = T,
  fix_indels = F,
  ...
)
}
\arguments{
\item{subject}{a named character or named DNAStringSet of one subject (only the DNAStringSet but not DNAString can hold a name)}

\item{patterns}{a named character vector or named DNAStringSet of patterns to align to the subject sequence}

\item{type}{the type of alignment passed to Biostrings::pairwiseAlignment; not every type may work well with this function (if there are overlapping ranges of the alignments to the subject for example)}

\item{perfect.matches.only}{filter patterns for those which match the subject without gaps, insertions or substitutions before pairwise alignment}

\item{order.patterns}{order pattern increasingly by alignment position (start)}

\item{attach.nt}{add the length of the string to the name on the axis}

\item{fix_indels}{in case of overlapping indels and shared subject ranges, cut respective patterns to avoid indels}

\item{...}{additional arguments to Biostrings::pairwiseAlignment apart from subject, pattern and type}
}
\value{
a list:
base.plot ggplot object of alignment shows patterns colored by nt,
match.plot ggplot object of alignment shows patterns colored match, mismatch, etc,
base.df = df and match.df are the respective data.frames used for plotting,
seq min.max.subject.position indicates the outer limits of all aligned patterns (min = start position of first aligned pattern, max = end position of the last aligned pattern)
}
\description{
This function is useful plot the alignment position of multiple patterns in one subject.
It uses Biostrings::pairwiseAlignment(), obtains the individual alignment boundaries and converts
the results to a ggplot object with a few optional complementing information (see arguments).
The function will fail if a gap is induced in the subject and at least two pattern alignments overlap at this gap.
}
\examples{
\dontrun{
s <- stats::setNames("AAAACCCCTTTTGGGGAACCTTCC", "sub")
s <- Biostrings::DNAStringSet(s)
p <- stats::setNames(c("TTCC", "CCCC", "TTTT", "GGGG", "AAAA"), c("pat1", "pat2", "pat3", "pat4", "pat5"))
p <- Biostrings::DNAStringSet(p)
als <- igsc::MultiplePairwiseAlignmentsToOneSubject(subject = s, patterns = p, tile.border.color = "black")
als_ordered <- igsc::MultiplePairwiseAlignmentsToOneSubject(subject = s, patterns = p, tile.border.color = "black", order.patterns = T)
}
}
