% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleSequenceAlignmentDECIPHER.R
\name{MultipleSequenceAlignmentDECIPHER}
\alias{MultipleSequenceAlignmentDECIPHER}
\title{Have a Multiple Sequence Alignment calculated with DECIPHER, plotted as ggplot2 object}
\usage{
MultipleSequenceAlignmentDECIPHER(
  input.set,
  add.consensus.seq = T,
  print.disamb.cons.seq.border = T,
  collapse.duplicate.seqs = T,
  open.browser = F
)
}
\arguments{
\item{input.set}{Vector of DNA sequences as character or DNAStringSet()}

\item{add.consensus.seq}{logicle of a consensus sequence should be added to the alignment}

\item{print.disamb.cons.seq.border}{logicle if to print the borders of longest stretch of disambiguate consensus sequence}

\item{collapse.duplicate.seqs}{collapse identical sequences in the input.set to one sequence before alignemnt}

\item{open.browser}{open the browser via DECIPHERs command DECIPHER::BrowseSeqs in addition to having the plot returned}
}
\value{
a ggplot object of the multiple sequence alignment
}
\description{
Have a Multiple Sequence Alignment calculated with DECIPHER, plotted as ggplot2 object
}
\examples{
\dontrun{

}
}
