% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gtf.R
\name{plot_gtf}
\alias{plot_gtf}
\title{Plot ranges from a GTF file to see genome organization}
\usage{
plot_gtf(gtf_df, y = "transcript_id", color = c("strand", "overlapping"))
}
\arguments{
\item{gtf_df}{data frame from igsc::read_gtf}

\item{y}{y axis column}

\item{color}{color column; overlapping is computed internally}
}
\value{
list of ggplot and data frame
}
\description{
Plot ranges from a GTF file to see genome organization
}
\examples{
\dontrun{
gtf <- igsc::read_gtf(gtf_path)$gtf |>
  dplyr::filter(feature == "CDS")
plot_gtf(gtf, color = "overlapping")
}
}
