% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sc_reads.R
\name{plot_sc_reads}
\alias{plot_sc_reads}
\title{Title}
\usage{
plot_sc_reads(
  gtf_file,
  genome_file,
  bam_file,
  seqname = "chr2",
  gene = "CD8A",
  n_reads_sample = 40
)
}
\arguments{
\item{n_reads_sample}{}
}
\description{
Title
}
