% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printPairwiseAlignment.R
\name{printPairwiseAlignment}
\alias{printPairwiseAlignment}
\title{Print pairwise alignments of DNA or AA to console or txt file}
\usage{
printPairwiseAlignment(
  alignments,
  linewidth = 50,
  match_dots = NULL,
  print_pos = T,
  print_pos_end = F,
  use_align_starts = T,
  out_file = NULL
)
}
\arguments{
\item{out_file}{}
}
\value{

}
\description{
Print pairwise alignments of DNA or AA to console or txt file
}
