% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwalign_print.R
\name{pwalign_print}
\alias{pwalign_print}
\title{Print pairwise alignments of DNA or AA to console or txt file}
\usage{
pwalign_print(
  pa,
  linewidth = 100,
  match_dots = NULL,
  print_pos = T,
  print_pos_end = F,
  use_align_starts = T,
  col_out = T,
  extend_subject = c(0, 0),
  out_file = NULL
)
}
\arguments{
\item{pa}{pwalign::pariwiseAlignment or XStringSet}

\item{linewidth}{numeric, how many letters to print per line}

\item{match_dots}{should matching position be printed as dots?;
'subject' to have dots in the subject, 'pattern' to
have them in the pattern}

\item{print_pos}{print the positions of pattern and subject above and below
respectively}

\item{print_pos_end}{print positions on the end of each line}

\item{use_align_starts}{if TRUE, the first positions are not 1 but refer to
position of alignment within the provided sequences}

\item{col_out}{color the printed alignment?}

\item{extend_subject}{extend subject further than alignment limits (-,+)}

\item{out_file}{path to a file where to print results to; if NULL results are
printed in console}
}
\value{
alignment in printed format in console or file
}
\description{
Print pairwise alignments of DNA or AA to console or txt file
}
\examples{
granzymes <- c("GZMA","GZMB","GZMH","GZMK","GZMM")
out <- get_sequences_from_biomart(granzymes)

gzma <- dplyr::filter(out, hgnc_symbol == "GZMA")
gzmb <- dplyr::filter(out, hgnc_symbol == "GZMB")
GZMA_CDS <- stats::setNames(gzma$coding, "GZMA_CDS")
GZMB_CDS <- stats::setNames(gzmb$coding, "GZMB_CDS")
aln6 <- pwalign::pairwiseAlignment(subject = Biostrings::DNAStringSet(GZMB_CDS),
                                   pattern = Biostrings::DNAStringSet(GZMA_CDS),
                                   type = "global")
pwalign_print(pwalign_to_xstringset(aln6), match_dots = "subject", col_out = T)
pwalign_print(aln6, match_dots = "subject", col_out = T)
}
