% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwalign_to_df.R
\name{pwalign_to_df}
\alias{pwalign_to_df}
\title{Convert pairwiseAlignment object to long data frame}
\usage{
pwalign_to_df(pa, verbose = T, subject_width = c("whole", "aligned"))
}
\arguments{
\item{pa}{pairwiseAlignment from pwalign::pairwiseAlignment}

\item{verbose}{print messages?}

\item{subject_width}{include whole subject width or pattern-aligned range only}
}
\value{
long data frame
}
\description{
Wrapper around pwalign_to_xstringset and xstringset_to_df.
df may be used for plotting. Or passed to compare_seq_df_long.
}
