% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwalign_to_df.R
\name{pwalign_to_xstringset}
\alias{pwalign_to_xstringset}
\title{Convert pairwiseAlignment object to XStringSet}
\usage{
pwalign_to_xstringset(pa, verbose = T, subject_width = c("whole", "aligned"))
}
\arguments{
\item{pa}{pairwiseAlignment from pwalign::pairwiseAlignment}

\item{verbose}{print messages?}

\item{subject_width}{include whole subject width or pattern-aligned range only}
}
\value{
XStringSet
}
\description{
E.g. for printing. See DECIPHER::BrowseSeqs.
}
