% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cellranger_outs.R
\name{read_cellranger_outs}
\alias{read_cellranger_outs}
\title{Read necessary file from 10X Genomics' cellranger output and return a concatenated data frame}
\usage{
read_cellranger_outs(vdj_path)
}
\arguments{
\item{vdj_path}{path to the parent folder containing one or multiple subfolders with cellrangers VDJ outs}
}
\value{
data frame (cl_long)
}
\description{
Required files in each subfolder of vdj_path are consensus_annotations.csv, filtered_contig_annotations.csv, consensus.fasta, concat_ref.fasta, filtered_contig.fasta.
If one or more are missing this subfolder is excluded. clonotype_id and consensus_id are suffixed with _cr for cellranger; entries from consensus.fasta becomes consensus_seq_cr.
...
}
