% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_paired_reads.R
\name{read_paired_reads}
\alias{read_paired_reads}
\title{Read paired reads from RNA sequencing into memory}
\usage{
read_paired_reads(
  fastq_path_r1,
  fastq_path_r2,
  min_len = NULL,
  filter_ACTG = T,
  filter_paired = F,
  quality_filter = T,
  max_freq_below_Q30 = 0.1,
  vroom_lines_args = list(progress = F, skip_empty_rows = T),
  sample = 1,
  mc.cores = 1
)
}
\arguments{
\item{fastq_path_r1}{path to .fastq or .fastq.gz file of read1}

\item{fastq_path_r2}{path to .fastq or .fastq.gz file of read2}

\item{min_len}{minimal length of reads to return}

\item{filter_ACTG}{filter for reads with letter ACTG only; remove all reads which contain N for instance}

\item{filter_paired}{if one of paired reads does not pass a test (min_len, filter_ACTG, quality_filter, max_freq_below_Q30),
then remove both reads (TRUE) or only the respective read (FALSE) by setting it to NA}

\item{quality_filter}{perform quality filtering on reads according to max_freq_below_Q30}

\item{max_freq_below_Q30}{maximum frequency of bases per reads that are allowed to be below Q30}

\item{vroom_lines_args}{arguments to vroom::lines}

\item{sample}{return a subset (sample) of paired reads only; can be a value greater
than 1 to sample a fixed absolute number (limited to total number of reads) or
a value lower than 1 to sample a fraction of reads}

\item{mc.cores}{number of cores to use to parallel computing, speeds up quality filtering;
is limited to parallel::detectCores()}
}
\value{
a list with read sequences, their quality strings and statistics
}
\description{
Reads fastq files from Illumina paired sequencing and applied optional filtering
by quality metrics.
}
\examples{
\dontrun{
  reads_paired <- igsc::read_paired_reads(
    fastq_path_r1 = r1_path,
    fastq_path_r2 = r2_path,
    vroom_lines_args = list(progress = T,
                            skip_empty_rows = T),
    min_len = 40,
    mc.cores = 8)
}
}
