% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_seq.R
\name{rotate_seq}
\alias{rotate_seq}
\title{Rotate character strings by splitting at a given position}
\usage{
rotate_seq(seq, cut = 1)
}
\arguments{
\item{seq}{A character vector where each element is a string to be rotated.}

\item{cut}{A single integer between 1 and the length of each string in \code{seq} (inclusive). Defaults to 1.}
}
\value{
A character vector of the same length and names as \code{seq}, with each string rotated.
}
\description{
This function rotates each string in a character vector by splitting at position \code{cut} (1-indexed) and swapping the two parts.
For example, "abcdef" with \code{cut = 3} becomes "cdefab".
}
\details{
The function will stop if \code{cut} is less than 1 or greater than the length of any string in \code{seq}.
}
\examples{
# Basic rotation
rotate_seq("abcdef", cut = 3)
# [1] "cdefab"

# Vector of strings
rotate_seq(c("abc", "defg", "hijkl"), cut = 2)
# [1] "cabc" "fgde" "ijkhl"

# Named input preserves names
rotate_seq(c(a = "abcdef", b = "ghijkl"), cut = 3)
#        a        b
# "cdefab" "ijklgh"

}
