% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdjdb_hits.R
\name{vdj_hits}
\alias{vdj_hits}
\title{Title}
\usage{
vdj_hits(
  vdjdb,
  tcrs,
  vdj_tr_col = "Gene",
  tcr_tr_col = "chain",
  vdj_cdr3_col = "CDR3",
  tcr_cdr3_col = "CDR3_aa_cr",
  separate_TRA_TRB_only = T,
  ...
)
}
\arguments{
\item{vdjdb}{data.frame of TCRs from vdjdb, first column indicating the chain TRA or TRB, second the aa sequence of CDR3}

\item{tcrs}{data.frame of TCRs from sequencing, first column indicating the chain TRA or TRB, second the aa sequence of CDR3}
}
\value{

}
\description{
Title
}
