% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algnmt_plot.R
\name{xstringset_to_df}
\alias{xstringset_to_df}
\title{Convert XStringSet to data frame}
\usage{
xstringset_to_df(
  xstringset,
  name_col = "seq.name",
  seq_col = "seq",
  pos_col = "position",
  subject_name = NULL,
  format = c("long", "wide")
)
}
\arguments{
\item{xstringset}{AA/DNA/RNA StringSet from Biostrings::}

\item{name_col}{column with seq names}

\item{seq_col}{column with seq data (nt or aa)}

\item{pos_col}{position column}

\item{subject_name}{name of sequence that is subject or reference;
is passed as attribute to return df}

\item{format}{join sequences to long or wide df?}
}
\value{
data frame
}
\description{
Convert XStringSet to data frame
}
