% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_imgt_ref_to_TCR_seq.R
\name{align_imgt_ref_to_TCR_seq}
\alias{align_imgt_ref_to_TCR_seq}
\title{Align IMGTs reference sequences to your TCR sequence}
\usage{
align_imgt_ref_to_TCR_seq(
  chain,
  TCR,
  cl_long,
  cl_wide,
  imgt_ref,
  sequence_col = "consensus_seq_cr",
  C_allele,
  type = "local",
  ...
)
}
\arguments{
\item{chain}{character which TCR chain to align, currently only TRA and TRB}

\item{TCR}{a named vector; which of your TCRs to align, the name indicates which column name of cl_long to use, the value indicates what to look for in that column}

\item{cl_long}{clonotype data frame long format}

\item{cl_wide}{clonotype data frame wide format}

\item{imgt_ref}{the prepared data.frame of IMGT references}

\item{sequence_col}{name of the column to pull sequences from}

\item{C_allele}{optional name of the constant allele to use for the alignment, must be the entry in the "Allele"-column of the imgt_ref data frame}
}
\value{
a list
}
\description{
With Biostrings::pairwiseAlignment() the borders of the V(D)J-segments in your TCR sequence can be determined.
The sequence within these border may then be used for further steps. In case of mouse, IMGT gene segments from Mus musculus are preferred.
}
\examples{
\dontrun{

}
}
