% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_reads.R
\name{align_reads}
\alias{align_reads}
\title{Align reads r1 and r2 to a set of reference sequences}
\usage{
align_reads(
  r1,
  r2,
  r1_table = NULL,
  r2_table = NULL,
  ref_seq_list,
  mapply_fun = parallel::mcmapply,
  maxmis = 0,
  ...
)
}
\arguments{
\item{r1}{vector of r1 reads from paired sequencing, must be same length as r2}

\item{r2}{see r1}

\item{r1_table}{optional, table of r1 generated with table(r1, useNA = "no")}

\item{r2_table}{see r1_table}

\item{ref_seq_list}{names list of vectors of reference seqences to align r1 and r2 to}

\item{...}{}
}
\description{
Align reads r1 and r2 to a set of reference sequences
}
\examples{
\dontrun{
reads_paired <- igsc::read_paired_reads(fastq_path_r1 = r1_path, fastq_path_r2 = r2_path, min_len = 40, mc.cores = 8)
r1_table <- table(reads_paired[["reads"]][["r1"]][["seq"]], useNA = "no")
r2_table <- table(reads_paired[["reads"]][["r2"]][["seq"]], useNA = "no")
read_align_list <- align_reads(r1 = reads_paired[["reads"]][["r1"]][["seq"]],
  r2 = reads_paired[["reads"]][["r2"]][["seq"]],
  r1_table = r1_table,
  r2_table = r2_table,
  ref_seq_list = list(ref1 = c("ATCGTGCTG", "ACTTTTGGGCTTAAA"), ref2 = c("AAAGGTTTCCCGTC", "TTGGCCAA", "GGCTA")),
  mc.cores = 2)
}
}
