% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_imgt_allele.R
\name{attach_imgt_alleles}
\alias{attach_imgt_alleles}
\title{Match best matching IMGT reference alleles (TCR V and J) and attache respective columns to cl_long}
\usage{
attach_imgt_alleles(
  cl_long,
  imgt_ref,
  pick.by = "alignment",
  lapply_fun = lapply,
  ...
)
}
\arguments{
\item{cl_long}{clonotype data in long format, preferentially from igsc::read_cellranger_outs}

\item{imgt_ref}{imgt reference data frame}

\item{pick.by}{match disambiguate alleles from IMGT by best match (alignment) or just randomly (random);
random is offered to speed up the process in case of many alleles}

\item{lapply_fun}{function name without quotes; lapply, pbapply::pblapply or parallel::mclapply are suggested}

\item{...}{additional argument to the lapply function; mainly mc.cores when parallel::mclapply is chosen}
}
\value{

}
\description{
Match best matching IMGT reference alleles (TCR V and J) and attache respective columns to cl_long
}
