% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_seq_df_long.R
\name{compare_seq_df_long}
\alias{compare_seq_df_long}
\title{Title}
\usage{
compare_seq_df_long(
  df_long,
  ref,
  pos_col = "position",
  seq_col = "seq",
  name_col = "seq.name",
  seq_original = "seq_original",
  match_symbol = ".",
  mismatch_symbol = "x",
  change_pattern = F,
  pattern_mismatch_as = c("base", "mismatch_symbol"),
  change_ref = F,
  ref_mismatch_as = c("base", "mismatch_symbol"),
  insertion_as = c("base", "insertion")
)
}
\arguments{
\item{insertion_as}{}
}
\description{
Title
}
