% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_transcript.R
\name{concat_transcript}
\alias{concat_transcript}
\title{Pull and concatenate exon ranges into one sequence}
\usage{
concat_transcript(
  gtf_df,
  refseq,
  refseq_strand = c("+", "-"),
  stop_codon_to_CDS = F,
  run_test = F,
  ...
)
}
\arguments{
\item{gtf_df}{subsetted gtf data frame}

\item{refseq}{reference sequence in 5' to 3' direction, genomic sequence: plus or minus strand}

\item{refseq_strand}{which strand does refseq represent?}

\item{stop_codon_to_CDS}{include stop codon in CDS?}

\item{run_test}{compare returned sequences to references?}

\item{...}{arguments to get_genome_seq}
}
\description{
Based on gtf_df which should only contain lines of exon definitions for
one gene, the respective sequences are pulled from refseq and concatenated
into the coding sequence (CDS).
Stop codon added to CDS - add option
}
\details{
https://www.biostars.org/p/3423/, https://pmc.ncbi.nlm.nih.gov/articles/PMC6099125/, https://en.wikipedia.org/wiki/Sense_(molecular_biology)
}
