% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwalign_to_df.R
\name{df_to_xstringset}
\alias{df_to_xstringset}
\title{Title}
\usage{
df_to_xstringset(
  df,
  type = c("DNA", "RNA", "AA"),
  name_col = "seq.name",
  seq_col = "seq",
  sym_internal_gap = "-",
  sym_terminal_gap = "-"
)
}
\arguments{
\item{sym_terminal_gap}{}
}
\description{
Title
}
