% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_fasta_to_fst.R
\name{genome_fasta_to_fst}
\alias{genome_fasta_to_fst}
\title{Title}
\usage{
genome_fasta_to_fst(
  genome_file,
  subfolder = "genome_fagz",
  overwrite = F,
  compression = 100,
  verbose = T,
  simplify_names = T
)
}
\arguments{
\item{subfolder}{"genome_fst"}

\item{simplify_names}{}
}
\description{
Title
}
