% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fasta_seq_bounds.R
\name{get_fasta_seq_bounds}
\alias{get_fasta_seq_bounds}
\title{From a fasta file, get sequence boundaries (lines)}
\usage{
get_fasta_seq_bounds(file_path)
}
\arguments{
\item{file_path}{path to fasta file or fst file}

\item{overwrite_fst}{}
}
\description{
Use forcats::fct_drop to drop unused factor levels.
}
