% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_sequences.R
\name{get_gene_sequences}
\alias{get_gene_sequences}
\title{Title}
\usage{
get_gene_sequences(gtf_path, genome_path, gene_name)
}
\arguments{
\item{gene_name}{}
}
\description{
Title
}
\examples{
\dontrun{
out <- get_gene_sequences(gtf_path = "/Users/chris/Documents/2024_igsc_testing/refdata-gex-GRCh38-2020-A/genes.gtf",
                          genome_path = "/Users/chris/Documents/2024_igsc_testing/refdata-gex-GRCh38-2020-A/genome.fa",
                          gene_name = "GZMB")
}
}
