% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_seq.R
\name{get_genome_seq}
\alias{get_genome_seq}
\title{Title}
\usage{
get_genome_seq(
  fasta_file = NULL,
  fst_folder = NULL,
  fagz_folder = NULL,
  ucsc_api = "http://api.genome.ucsc.edu/getData/sequence?genome=hg38&",
  query_string = NULL,
  chromosome = NULL,
  start = NULL,
  end = NULL,
  ucsc_toupper = T
)
}
\arguments{
\item{fagz_folder}{}
}
\description{
Title
}
