% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sequences_from_biomart.R
\name{get_sequences_from_biomart}
\alias{get_sequences_from_biomart}
\title{Get gene, transcript, amino acid sequences from biomaRt}
\usage{
get_sequences_from_biomart(hgnc_symbol, dataset = "hsapiens_gene_ensembl")
}
\arguments{
\item{hgnc_symbol}{gene symbol}

\item{dataset}{passed to biomaRt::useEnsembl}
}
\value{
data frame
}
\description{
Retrieve canonical sequences. coding = cds (exons), cDNA = exons + UTR,
transcript_exon_intron = introns+exons, gene_exon_intron = full genomic span,
gene_exon = exons in genomic order, not revcomp
}
\examples{
granzymes <- c("GZMA","GZMB","GZMH","GZMK","GZMM")
out <- get_sequences_from_biomart(granzymes)
}
