% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_align.R
\name{global_align}
\alias{global_align}
\title{Title}
\usage{
global_align(
  pattern,
  subject,
  semi_global = T,
  match = 1,
  mismatch = -1,
  gap = -1
)
}
\arguments{
\item{gap}{}
}
\description{
Title
}
