% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_typing_from_bam.R
\name{hla_typing_from_bam}
\alias{hla_typing_from_bam}
\title{Type HLA genes directly from an RNA-seq BAM file}
\usage{
hla_typing_from_bam(
  bam,
  hla_ref,
  reference_genome = NULL,
  reference_gtf = NULL,
  regions = NULL,
  genes = c("A", "B", "C"),
  hla_gene_col_name = "gene",
  mhc_start = 28000000L,
  mhc_end = 34000000L,
  include_alt_contigs = TRUE,
  tags = c("CR", "CB", "CY", "AS", "UR", "UB", "UY", "HI", "NH", "nM", "RE"),
  scores = TRUE,
  min_mapq = 0L,
  primary_only = TRUE,
  cell_barcodes = NULL,
  cell_barcode_tag = "CB",
  allele_diff = 5,
  top_n_pairwise_results = 50,
  hla_seq_col_name = "seq_Exon2_3",
  hla_allele_col_name = "allele",
  p_group_col_name = "p_group",
  g_group_col_name = "g_group",
  lapply_fun = lapply,
  maxmis = 3,
  make_reads_distinct = FALSE,
  ...
)
}
\arguments{
\item{bam}{Path to a coordinate-sorted BAM file with a corresponding index.}

\item{hla_ref}{HLA allele reference data frame, preferably created by
\code{\link[=hla_df_from_xml]{hla_df_from_xml()}}.}

\item{reference_genome}{Optional path to the FASTA file used for alignment.
The FASTA must have a \code{.fai} index and is used to validate reference names
and lengths.}

\item{reference_gtf}{Optional path to a reference GTF file. When \code{regions}
is \code{NULL}, gene records are read with \code{\link[=read_gtf]{read_gtf()}} and the annotated
boundaries of \code{genes} are used as the query ranges.}

\item{regions}{Optional \code{\link[GenomicRanges:GRanges]{GenomicRanges::GRanges()}} object. Explicit regions
take precedence over \code{reference_gtf}. When both are \code{NULL}, the chromosome
6 interval specified by \code{mhc_start} and \code{mhc_end} is used and, when
\code{include_alt_contigs = TRUE}, chromosome 6 alternative and HLA-named
contigs are added.}

\item{genes}{Character vector of genes to type, for example \code{c("A", "B", "C")}. An optional \verb{HLA-} prefix is ignored. Use \code{NULL} to type every gene
represented in \code{hla_ref}. When \code{reference_gtf} is supplied, the same genes
determine which annotated gene boundaries are queried.}

\item{hla_gene_col_name}{Column in \code{hla_ref} containing HLA gene names.}

\item{mhc_start, mhc_end}{One-based inclusive bounds of the broad MHC interval
used when \code{regions} is \code{NULL}.}

\item{include_alt_contigs}{Logical; include whole BAM reference sequences
whose names identify them as chromosome 6 alternative or HLA contigs.}

\item{tags}{BAM tags passed to \code{\link[=get_bam_reads]{get_bam_reads()}}. The default includes common
Cell Ranger cell-barcode, UMI, and alignment tags.}

\item{scores}{Logical; calculate base-quality summaries in
\code{\link[=get_bam_reads]{get_bam_reads()}}.}

\item{min_mapq}{Minimum mapping quality to retain.}

\item{primary_only}{Logical; remove secondary and supplementary alignments.}

\item{cell_barcodes}{Optional character vector of corrected cell barcodes to
retain. This requires the tag named by \code{cell_barcode_tag}.}

\item{cell_barcode_tag}{Name of the BAM tag containing corrected cell
barcodes.}

\item{allele_diff}{Maximum fold difference used to retain candidate alleles;
passed to \code{\link[=hla_typing]{hla_typing()}}.}

\item{top_n_pairwise_results}{Number of leading allele pairs to plot; passed
to \code{\link[=hla_typing]{hla_typing()}}.}

\item{hla_seq_col_name}{Name of the HLA sequence column in \code{hla_ref}.}

\item{hla_allele_col_name}{Name of the allele identifier column in
\code{hla_ref}.}

\item{p_group_col_name}{Name of the P-group column in \code{hla_ref}.}

\item{g_group_col_name}{Name of the G-group column in \code{hla_ref}.}

\item{lapply_fun}{Apply function used for allele matching; passed to
\code{\link[=hla_typing]{hla_typing()}}.}

\item{maxmis}{Maximum mismatches allowed per read match; passed to
\code{\link[=hla_typing]{hla_typing()}}.}

\item{make_reads_distinct}{Logical; remove duplicate read sequences before
matching; passed to \code{\link[=hla_typing]{hla_typing()}}.}

\item{...}{Additional arguments passed to \code{lapply_fun} by \code{\link[=hla_typing]{hla_typing()}},
such as \code{mc.cores} for \code{\link[parallel:mclapply]{parallel::mclapply()}}.}
}
\value{
A list with \code{reads}, the candidate HLA read data frame; \code{typing}, a
named list of \code{\link[=hla_typing]{hla_typing()}} results; and \code{regions}, the queried genomic
ranges.
}
\description{
A convenience workflow that first calls \code{\link[=get_hla_reads]{get_hla_reads()}} and then calls
\code{\link[=hla_typing]{hla_typing()}} separately for each requested HLA gene.
}
\examples{
\dontrun{
result <- hla_typing_from_bam(
  bam = "outs/possorted_genome_bam.bam",
  hla_ref = hla_ref,
  reference_gtf = "reference/genes/genes.gtf",
  genes = c("A", "B", "C"),
  maxmis = 1)

# check strandness of HLA genes
gtf <- read_gtf(".../refdata-gex-GRCh38-2020-A/genes/genes.gtf.gz",
                gene_names = "HLA-",
                gene_names_full_match = F,
                features = "gene")$gtf |>
  dplyr::select(gene_name, start, end, strand) |>
  dplyr::filter(!grepl("AS", gene_name)) |>
  dplyr::arrange(gene_name)
# genes on (-) Strand: revcomp the reference as reads from BAM file are always
# mapped to (+) Strand but reference give sequence from (-) Strand
#   |gene_name |    start|      end|strand |
#   |:---------|--------:|--------:|:------|
#   |HLA-A     | 29941260| 29945884|+      |
#   |HLA-B     | 31269491| 31357188|-      |
#   |HLA-C     | 31268749| 31272130|-      |
#   |HLA-DMA   | 32948613| 32969094|-      |
#   |HLA-DMB   | 32934629| 32941028|-      |
#   |HLA-DOA   | 33004182| 33009591|-      |
#   |HLA-DOB   | 32812763| 32820466|-      |
#   |HLA-DPA1  | 33064569| 33080775|-      |
#   |HLA-DPB1  | 33075990| 33089696|+      |
#   |HLA-DQA1  | 32628179| 32647062|+      |
#   |HLA-DQA2  | 32741391| 32747198|+      |
#   |HLA-DQB1  | 32659467| 32668383|-      |
#   |HLA-DQB2  | 32756098| 32763532|-      |
#   |HLA-DRA   | 32439878| 32445046|+      |
#   |HLA-DRB1  | 32578769| 32589848|-      |
#   |HLA-DRB5  | 32517353| 32530287|-      |
#   |HLA-E     | 30489509| 30494194|+      |
#   |HLA-F     | 29722775| 29738528|+      |
#   |HLA-G     | 29826967| 29831125|+      |


library(igsc)
hla <- hla_df_from_xml(
  "/Volumes/CMS_SSD_2TB/hla.xml.gz",
  lapply_fun = parallel::mclapply,
  mc.cores = 8)

hlaA <- hla |>
  dplyr::filter(grepl("^HLA-A", allele)) |>
  dplyr::filter(seq_length == max(seq_length), .by = allele_protein)

hlaresA <- hla_typing_from_bam(
  bam = "/Users/chris/Documents/possorted_genome_bam.bam",
  reference_gtf = "/Volumes/CMS_SSD_2TB/reference_genomes/refdata-gex-GRCh38-2020-A/genes/genes.gtf.gz",
  hla_ref = hlaA,
  genes = "A",
  lapply_fun = parallel::mclapply,
  make_reads_distinct = T,
  mc.cores = 4,
  maxmis = 3,
  allele_diff = 4)
hlaresA$typing$B$plot_pair_res2

## HLA-B and HLA-C are on (-) Strand but all bam reads are returned as on (+) Strand
## hence revcomp the reference: (-) --> (+)
hlaB <- hla |>
  dplyr::filter(grepl("^HLA-B", allele)) |>
  dplyr::filter(seq_length == max(seq_length), .by = allele_protein) |>
  dplyr::mutate(seq_Exon2_3 = revcompDNA(seq_Exon2_3))

hlaresB <- hla_typing_from_bam(
  bam = "/Users/chris/Documents/possorted_genome_bam.bam",
  reference_gtf = "/Volumes/CMS_SSD_2TB/reference_genomes/refdata-gex-GRCh38-2020-A/genes/genes.gtf.gz",
  hla_ref = hlaB,
  genes = "B",
  lapply_fun = parallel::mclapply,
  make_reads_distinct = T,
  mc.cores = 4,
  maxmis = 3,
  allele_diff = 4)
hlaresB$typing$B$plot_pair_res2

hlaC <- hla |>
  dplyr::filter(grepl("^HLA-C", allele)) |>
  dplyr::filter(seq_length == max(seq_length), .by = allele_protein) |>
  dplyr::mutate(seq_Exon2_3 = revcompDNA(seq_Exon2_3))

hlaresC <- hla_typing_from_bam(
  bam = "/Users/chris/Documents/possorted_genome_bam.bam",
  reference_gtf = "/Volumes/CMS_SSD_2TB/reference_genomes/refdata-gex-GRCh38-2020-A/genes/genes.gtf.gz",
  hla_ref = hlaC,
  genes = "C",
  lapply_fun = parallel::mclapply,
  make_reads_distinct = T,
  mc.cores = 4,
  maxmis = 3,
  allele_diff = 4)
hlaresC$typing$C$plot_pair_res2

hlaDPA1 <- hla |>
  dplyr::filter(grepl("^HLA-DPA1", allele)) |>
  dplyr::filter(seq_length == max(seq_length), .by = allele_protein) |>
  dplyr::mutate(seq_Exon2_3 = revcompDNA(seq_Exon2_3))

hlaresDPA1 <- hla_typing_from_bam(
  bam = "/Users/chris/Documents/possorted_genome_bam.bam",
  reference_gtf = "/Volumes/CMS_SSD_2TB/reference_genomes/refdata-gex-GRCh38-2020-A/genes/genes.gtf.gz",
  hla_ref = hlaDPA1,
  genes = "DPA1",
  lapply_fun = parallel::mclapply,
  make_reads_distinct = T,
  mc.cores = 4,
  maxmis = 3,
  allele_diff = 2)
hlaresDPA1$typing$DPA1$plot_pair_res2
}
}
