% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiplePairwiseAlignmentsToOneSubject.R
\name{MultiplePairwiseAlignmentsToOneSubject}
\alias{MultiplePairwiseAlignmentsToOneSubject}
\title{Align multiple DNA pattern sequences to one subject}
\usage{
MultiplePairwiseAlignmentsToOneSubject(
  subject,
  patterns,
  subject.name,
  patterns.names,
  type = "global-local",
  subject.alignment.limits,
  print.pattern.positions = T,
  pattern.positions.size = 2,
  print.subject.min.max = F,
  attach.length.to.name = T,
  tile.border.color = NA,
  title
)
}
\arguments{
\item{subject}{a character or DNAStringSet of one subject (to provide a name in final graphics, provide a named DNAStringSet)}

\item{patterns}{a character vector or DNAStringSet of patterns to align to the subject sequence}

\item{subject.name}{optional, provide a name for the subject; overwrites names provided with subject}

\item{patterns.names}{optional, provide names for the patterns; overwrites names provided with patterns}

\item{type}{the type of alignment passed to Biostrings::pairwiseAlignment; not every type may work well with this function (if there are overlapping ranges of the alignments to the subject for example)}

\item{subject.alignment.limits}{numeric vector of nt-positions (lower and upper) of the subject to manually limit (cut) the final graphic}

\item{print.pattern.positions}{check}

\item{pattern.positions.size}{check}

\item{print.subject.min.max}{check}

\item{attach.length.to.name}{add the length of the string to the name on the axis}

\item{tile.border.color}{character; tiles from geom_tile are used to plot nts - should they have a border color, e.g. "black"; only useful for short alignment and only an aesthetic thing}

\item{title}{print a title to the alignment}
}
\value{
a list:
base.plot ggplot object of alignment shows patterns colored by nt,
match.plot ggplot object of alignment shows patterns colored match, mismatch, etc,
base.df = df and match.df are the respective data.frames used for plotting,
seq min.max.subject.position indicates the outer limits of all aligned patterns (min = start position of first aligned pattern, max = end position of the last aligned pattern)
}
\description{
This function is useful plot the alignment position of multiple patterns in one subject.
It uses Biostrings::pairwiseAlignment(), obtains the individual alignment limits and converts that to a ggplot object with a few optional complementing information.
The function will not work if a gap is induced in the subject and at least two patterns overlap at this gap.
}
\examples{
\dontrun{
s <- stats::setNames("AAAACCCCTTTTGGGGAACCTTCC", "sub")
s <- Biostrings::DNAStringSet(s)
p <- stats::setNames(c("AAAA", "CCCC", "TTTT", "GGGG", "TTCC"), c("pat1", "pat2", "pat3", "pat4", "pat5"))
p <- Biostrings::DNAStringSet(p)
als <- igsc::MultiplePairwiseAlignmentsToOneSubject(subject = s, patterns = p)

}
}
