% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_randomNames.R
\name{pick_randomNames}
\alias{pick_randomNames}
\title{Pick n random names from the randomNames package}
\usage{
pick_randomNames(
  n,
  names_to_avoid = NULL,
  max_iter = 500,
  min_name_nchar = 3,
  avoid_chars = c(" ", "-", "'", ",", ";"),
  randomNames_args = list(which.names = "first")
)
}
\arguments{
\item{n}{numeric (integer) of how many unique names to return}

\item{names_to_avoid}{a vector of names to avoid}

\item{max_iter}{maximum number of while-loop iterations to avoid an infinite loop}

\item{min_name_nchar}{min length of name returned by randomNames}

\item{avoid_chars}{character vector of forbidden characters in names or NULL;
reject names with one or more of these characters}

\item{...}{arguments passed to randomNames::randomNames}
}
\value{
a vector of n unique random names
}
\description{
The randomNames function from randomNames does not necessarily provide unique names (https://github.com/CenterForAssessment/randomNames/issues/55).
Hence this function has been written to fix that issue as elegant as possible.
}
