% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printPairwiseAlignment.R
\name{printPairwiseAlignment}
\alias{printPairwiseAlignment}
\title{Print pairwise alignments of DNA or AA to console or txt file}
\usage{
printPairwiseAlignment(
  alignments,
  linewidth = 50,
  match_dots = NULL,
  print_pos = T,
  print_pos_end = F,
  use_align_starts = T,
  col_out = T,
  extend_subject = c(0, 0),
  out_file = NULL
)
}
\arguments{
\item{alignments}{one alignment or a list of multiple alignments created with
Biostrings::pariwiseAlignment}

\item{linewidth}{numeric, how many letters of alignment to print per line}

\item{match_dots}{should matching position be printed as dots for easier
recognition of differences; 'subject' to have dots in the subject, 'pattern' to
have them in the pattern}

\item{print_pos}{print the positions of pattern and subject above and below
respectively}

\item{print_pos_end}{print positions on the end of each line}

\item{use_align_starts}{if TRUE, the first positions are not 1 but refer to
position of alignment within the provided sequences}

\item{col_out}{color the printed alignment?}

\item{extend_subject}{extend subject further than alignment limits (-,+)}

\item{out_file}{path to a file where to print results to; if NULL results are
printed in console}
}
\value{
alignment in printed format in console or file
}
\description{
Print pairwise alignments of DNA or AA to console or txt file
}
