% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gtf.R
\name{process_gtf_attribute_col}
\alias{process_gtf_attribute_col}
\title{Process attribute column of gtf file}
\usage{
process_gtf_attribute_col(
  gtf,
  attr_keep = NULL,
  attr_rename = NULL,
  rename_replace = F,
  attr_as = c("cols", "kv"),
  gene_name_prefix = "",
  gene_name_force = "gene_id",
  gene_name_fix = T,
  gene_name_replace = list(),
  use_fun = "r",
  rm_index = T,
  rm_exon = F,
  exons_only = F,
  features_to_exon = c(""),
  aggregate_exons = F,
  aggregate_overlapping_exon_ranges = F,
  check_for_rotation = F,
  genome_length = NULL,
  rm_entries_wo_matching_exon = F,
  verbose = T
)
}
\arguments{
\item{gtf}{gtf file as read with read_gtf}

\item{attr_keep}{which attributes to keep from attribute column; NULL to keep all;
e.g. attr_keep = c("gene_id", "transcript_id", "gene_name", "gene_type")}

\item{attr_rename}{key value pairs how to rename attributes; will append to attr_keep;
e.g. attr_rename = c("product" = "gene_name", "gene_biotype" = "gene_type")}

\item{rename_replace}{remove old entries from attr_rename}

\item{attr_as}{how to return attributes; kv is key value pairs}

\item{gene_name_prefix}{prefix to add to gene_name?; e.g. "EBV"}

\item{gene_name_force}{if is.na(gene_name) which attribute to use instead for
gene_name; NULL to not do it; e.g. "gene_id"}

\item{gene_name_fix}{fix names: like replace all non-alphanumeric by dashes}

\item{gene_name_replace}{e.g. c("glycopeotein" = "")}

\item{use_fun}{which function to use for processing the attr_col; only r
is working and is fast}

\item{rm_index}{remove index column?}

\item{rm_exon}{remove original exon entries from gtf, done first}

\item{exons_only}{only retain exons entries; done third}

\item{features_to_exon}{features to become exons, e.g. "CDS"; done second}

\item{aggregate_exons}{aggregate exons of one transcript_id to one entry;
probably necessary for aggregate_overlapping_exon_ranges; information on intronic
regions lost; done fourth; may be relevant for mapping reads}

\item{aggregate_overlapping_exon_ranges}{check for overlap between transcripts;
when overlap: transcripts are joined to one entry; when both on same strand,
then this info is retained, when on different strands then (+)strand is assigned;
may be relevant for mapping reads}

\item{check_for_rotation}{check if any gene spans the artificial origin
(start) of a circular genome; if so, another cut position is chosen and
start end coordinates are adjusted; genome_length needed then}

\item{genome_length}{length of associated genome or refseq; only needed
for rotation}

\item{rm_entries_wo_matching_exon}{this has to done to meet mkref requirements}
}
\description{
Process attribute column of gtf file
}
\examples{
\dontrun{
# suggestion for viral genome prep as cellranger ref
gtf <- read_gtf(viral_gtf[[x]], process_attr_col = F)[["gtf"]]
process_gtf_attribute_col(gtf,
                          attr_keep = NULL,
                          attr_rename = y,
                          attr_as = "kv",
                          gene_name_prefix = paste0(x, "_"),
                          gene_name_force = "gene_id",
                          gene_name_replace = c("glycoprotein" = "",
                                                "helicase-primase primase subunit" = "",
                                                "helicase-primase subunit" = "",
                                                "single-stranded DNA binding protein" = "ssDNAbp",
                                                "hypothetical protein" = "HYPPROT"),
                          rm_index = T,
                          rm_exon = T,
                          exons_only = T,
                          aggregate_overlapping_exon_ranges = T,
                          aggregate_exons = T,
                          check_for_rotation = T,
                          genome_length = z,
                          features_to_exon = c("CDS"))
# then optional:
igsc:::ebvgenomerot <- rotate_genome_string(ebvgenome, cut = 144792)
igsc::write_fasta(ebvgenomerot, file = "ebv_rotated.fa")
# later:
combine_gtf_and_genome_for_cellranger(genome_files = viral_genomes[-3],
                                      gtf_files = viral_gtf_new_path,
                                      save_path = wd,
                                      save_names = c("genome5.fa", "genes5.gtf"),
                                      overwrite = T)
}
}
