% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwalign_multi.R
\name{pwalign_multi}
\alias{pwalign_multi}
\title{Separately align multiple pattern sequences to one subject}
\usage{
pwalign_multi(
  subject,
  patterns,
  type = c("global-local", "global", "local", "overlap", "local-global"),
  max_mismatch = NA,
  order_patterns = F,
  fix_subject_indels = F,
  rm_indel_inducing_pattern = F,
  seq_type = NULL,
  return_max_mismatch_info_only = F,
  compare_seq_df_long_args = list(seq_original = NULL, match_symbol = ".", change_nonref
    = T, nonref_mismatch_as = "base", change_ref = T, ref_mismatch_as = "base",
    insertion_as = "base"),
  pairwiseAlignment_args = list(),
  algnmt_plot_args = list(add_length_suffix = T, pattern_lim_size = 2, verbose = F),
  order_subject_ranges = F,
  verbose = T
)
}
\arguments{
\item{subject}{a named character or named DNAStringSet of one subject (only the DNAStringSet but not DNAString can hold a name)}

\item{patterns}{a named character vector or named DNAStringSet of patterns to align to the subject sequence}

\item{type}{the type of alignment passed to pwalign::pairwiseAlignment; not every type may work well with this function (if there are overlapping ranges of the alignments to the subject for example)}

\item{max_mismatch}{only use patterns that have a maximum number of mismatches
with the subject}

\item{order_patterns}{order pattern increasingly by alignment position (start)}

\item{fix_subject_indels}{in case of overlapping indels and shared subject ranges, cut respective patterns to avoid indels}

\item{rm_indel_inducing_pattern}{remove patterns that cause indels in subject
and hence may be a problem for plotting}

\item{seq_type}{set sequence type to AA or NT if necessary; if NULL
it is attempted to guess the type}

\item{return_max_mismatch_info_only}{only return information on mismatches of patterns with the subject;
in this case no alignment is calculated}

\item{compare_seq_df_long_args}{arguments to igsc::compare_seq_df_long}

\item{pairwiseAlignment_args}{arguments to pwalign::pairwiseAlignment}

\item{algnmt_plot_args}{arguments to igsc::algnmt_plot}

\item{verbose}{}
}
\value{
a list
}
\description{
This function is useful to visualise the alignment position of multiple patterns on one subject.
It uses pwalign::pairwiseAlignment(), obtains the individual alignment boundaries and converts
the results to a ggplot object.
The function will fail if a gap is induced in the subject and at least two pattern alignments overlap at this gap.
Method = local-global avoids indels in subject. Local cuts subject and pattern to the best matching sequence of both.
}
\examples{
s <- stats::setNames("AAAACCCCTTTTGGGGAACCTTCC", "sub")
s <- Biostrings::DNAStringSet(s)
p <- stats::setNames(c("TTCC", "CCCC", "TTTT", "GGGG", "AAAA"), c("pat1", "pat2", "pat3", "pat4", "pat5"))
p <- Biostrings::DNAStringSet(p)
als <- pwalign_multi(subject = s, patterns = p)
als_ordered <- pwalign_multi(subject = s, patterns = p, order_patterns = T,
                                                      compare_seq_df_long_args = list(seq_original = NULL,
                                                                                      match_symbol = ".",
                                                                                      change_nonref = T,
                                                                                      nonref_mismatch_as = "base",
                                                                                      change_ref = F,
                                                                                      ref_mismatch_as = "base",
                                                                                      insertion_as = "base"))
}
