% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cellranger_outs2.R
\name{read_cellranger_outs2}
\alias{read_cellranger_outs2}
\title{Read necessary file from 10X Genomics' cellranger output and return a concatenated data frame}
\usage{
read_cellranger_outs2(vdj_outs_path)
}
\arguments{
\item{vdj_outs_path}{named vector of paths to the outs-folder (or any folders containing the necessary file, specified above);
names will be added as "sample"-column to the output data frame}
}
\value{
data frame (cl_long)
}
\description{
Required files in each subfolder of vdj_outs_path are consensus_annotations.csv, filtered_contig_annotations.csv, consensus.fasta, concat_ref.fasta, filtered_contig.fasta.
If one or more are missing this subfolder is excluded. clonotype_id and consensus_id are suffixed with _cr for cellranger; entries from consensus.fasta becomes consensus_seq_cr.
...
}
