% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta.R
\name{read_fasta}
\alias{read_fasta}
\title{Read sequences from a fasta-formatted file}
\usage{
read_fasta(file, rm_comments = T, comment_indicator = c(";", "#"))
}
\arguments{
\item{file}{path to a file with fasta-formatted sequences}

\item{rm_comments}{logical whether to remove comment lines}

\item{comment_indicator}{first characters of a line which indicate a comment}
}
\value{
a named character vector of sequences in file
}
\description{
vroom::vroom_lines is used to quickly read lines of a text file (as a fasta file is one).
}
\examples{
\dontrun{
seqs <- igsc::read_fasta(file = "my/fasta/file.fasta")
}
}
