% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta.R
\name{read_fasta}
\alias{read_fasta}
\title{Read sequences from a fasta-formatted file}
\usage{
read_fasta(
  file,
  trimws = F,
  rm_comments = F,
  comment_indicator = c(";", "#"),
  rm_leading_arrow = T,
  make_names = F,
  make_names_fun = janitor::make_clean_names,
  make_names_args = list(),
  start_line = 1,
  end_line = Inf,
  concat = T,
  progress = T
)
}
\arguments{
\item{file}{path to a file with fasta-formatted sequences; file may
be gunzipped (ending with .gz)}

\item{trimws}{trim leading and trailing whitespaces? disable in case of trustworthy files. may take time for large fasta files}

\item{rm_comments}{remove comment lines? this is very slow large files, currently.}

\item{comment_indicator}{first characters of a line which indicate a comment}

\item{rm_leading_arrow}{remove leading arrow in fasta names?}

\item{make_names}{apply make_names to fasta names?}

\item{start_line}{passed to vroom::vroom_lines}

\item{end_line}{passed to vroom::vroom_lines to determine n_max}

\item{concat}{}
}
\value{
a named character vector of sequences
}
\description{
vroom::vroom_lines is used to quickly read lines of a text file (as a fasta file is one).
}
\examples{
\dontrun{
seqs <- igsc::read_fasta(file = "my/fasta/file.fasta")
seqs_zipped <- igsc::read_fasta(file = "my/fasta/file.fa.gz")
}
}
