% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gtf.R
\name{read_gtf}
\alias{read_gtf}
\title{Read and process a GTF (Gene Transfer Format) file}
\usage{
read_gtf(
  file_path,
  seqnames = NULL,
  features = NULL,
  gene_names = NULL,
  process_attr_col = T,
  process_attr_col_args = list(attr_keep = c("gene_id", "gene_name", "transcript_id",
    "transcript_name", "exon_number"), attr_rename = NULL, gene_name_prefix = "",
    gene_name_force = NULL, rm_index = T, use_fun = "r", rm_exon = F, exons_only = F,
    features_to_exon = c(""), aggregate_exons = F, aggregate_overlapping_exon_ranges = F,
    check_for_rotation = F, genome_length = NULL),
  col_names = c("seqname", "source", "feature", "start", "end", "score", "strand",
    "frame", "attribute")
)
}
\arguments{
\item{file_path}{path to the file; file may be gunzipped (ending with .gz)}

\item{seqnames}{seqnames to filter the gtf file for; will decrease computation time
required for processing the attribute column}

\item{features}{features to filter the gtf file for; will decrease computation time
required for processing the attribute column
c("gene", "transcript", "exon", "CDS", "start_codon", "stop_codon",
"five_prime_utr", "three_prime_utr", "Selenocysteine")}

\item{gene_names}{vector of gene names to filter for before attr col processing}

\item{process_attr_col}{convert the attribute column into separate columns}

\item{process_attr_col_args}{arguments to process_attr_col_args}

\item{col_names}{column names to assign to the gtf data frame;
changing seqname, feature or attribute will break this function;
better leave col_names as it is}
}
\value{
a list with (i) entries of the GTF file including the attribute
column as list and some attributes as separate columns and
(ii) the attributes as long data frame
}
\description{
Since GTF files come in an unhandy format, this function may help
to easily read them into memory. As a whole or partly by providing
seqnames and/or features. Processing the attribute column is computationally
costly. Start/end: In both cases, gene on plus or minus strand, end is larger
then start. This makes immediate sense for the plus strand but for the minus
strand, genes actually go from a larger to a smaller position (5'->3').
}
\details{
See https://www.ensembl.org/info/website/upload/gff.html?redirect=no for
explanation of GTF file format.
GTF files and genomic fasta files may be downloaded here https://www.ncbi.nlm.nih.gov/datasets/genome/
or here https://www.ensembl.org/index.html
}
