% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gtf.R
\name{read_gtf}
\alias{read_gtf}
\title{Read and process a GTF (Gene Transfer Format) file}
\usage{
read_gtf(file_path, ...)
}
\arguments{
\item{file_path}{path to the file}

\item{...}{arguments to vroom::vroom}
}
\value{
a list with (i) entries of the GTF file including the attribute column as list and some attributes as separate columns and (ii) the attributes as long data frame
}
\description{
Read and process a GTF (Gene Transfer Format) file
}
