% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gtf.R
\name{read_gtf}
\alias{read_gtf}
\title{Read and process a GTF (Gene Transfer Format) file}
\usage{
read_gtf(
  file_path,
  seqnames = NULL,
  features = NULL,
  process_attr_col = T,
  col_names = c("seqname", "source", "feature", "start", "end", "score", "strand",
    "frame", "attribute"),
  attr_keep = c("gene_id", "gene_name", "transcript_id", "transcript_name",
    "exon_number"),
  use_fun = c("r", "rust", "rcpp")
)
}
\arguments{
\item{file_path}{path to the file; file may be gunzipped (ending with .gz)}

\item{seqnames}{seqnames to filter the gtf file for; will decrease computation time
required for processing the attribute column}

\item{features}{features to filter the gtf file for; will decrease computation time
required for processing the attribute column}

\item{process_attr_col}{convert the attribute column into separate columns}

\item{col_names}{column names to assign to the gtf data frame;
changing seqname, feature or attribute will break this function;
better leave col_names as it is}

\item{attr_keep}{which attributes to keep from attribute column}

\item{use_fun}{which function to use for processing the attr_col; rcpp is
fastest currently}
}
\value{
a list with (i) entries of the GTF file including the attribute
column as list and some attributes as separate columns and
(ii) the attributes as long data frame
}
\description{
Since GTF files come in an unhandy format, this function may help
to easily read them into memory. As a whole or partly by providing
seqnames and/or features. Processing the attribute column is computationally
costly. Start/end: In both cases, gene on plus or minus strand, end is larger
then start. This makes immediate sense for the plus strand but for the minus
strand, genes actually go from a larger to a smaller position (5'->3').
}
\details{
See https://www.ensembl.org/info/website/upload/gff.html?redirect=no for
explanation of GTF file format.
GTF files and genomic fasta files may be downloaded here https://www.ncbi.nlm.nih.gov/datasets/genome/
or here https://www.ensembl.org/index.html
}
