% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrape_ncbi.R
\name{webscrape_ncbi}
\alias{webscrape_ncbi}
\title{Scrape data from NCBI nucleotide pages}
\usage{
webscrape_ncbi(accession)
}
\arguments{
\item{accession}{accession number to get data from}
}
\description{
Scrape data from NCBI nucleotide pages
}
