% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fasta.R
\name{write_fasta}
\alias{write_fasta}
\title{Write sequences to a file in fasta format}
\usage{
write_fasta(
  seqs,
  file,
  linewidth = 60,
  mc.cores = floor(parallel::detectCores()/4),
  gzip = F,
  verbose = T
)
}
\arguments{
\item{seqs}{named character vector or list of sequences; if a list only provide one sequence per index}

\item{file}{full path to the output file to be written; recommended file extension: .fa or .fasta}

\item{linewidth}{number of character per line (only valid for sequence, not names)}
}
\value{
no return; file written to disk
}
\description{
Write sequences to a file in fasta format
}
