% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fasta.R
\name{write_fasta}
\alias{write_fasta}
\title{Write sequences to a file in fasta format}
\usage{
write_fasta(seqs, file, mode = c("w", "a"), linewidth = 60)
}
\arguments{
\item{seqs}{named character vector or list of sequences; if a list only provide one sequence per index}

\item{file}{full path to the output file to be written; recommended file extension: .fa or .fasta}

\item{mode}{how to write file: 'w' to initiate or overwrite; 'a' to append an existing file}

\item{linewidth}{number of character per line (only valid for sequence, not names)}
}
\value{
no return; file written to disk
}
\description{
Write sequences to a file in fasta format
}
