% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gtf.R
\name{write_gtf}
\alias{write_gtf}
\title{Write a gtf data frames to file}
\usage{
write_gtf(gtf_df, file, header = "##gtf file")
}
\arguments{
\item{gtf_df}{gtf data frame}

\item{file}{file path where to save}

\item{header}{header lines}
}
\description{
Write a gtf data frames to file
}
