% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gtf.R
\name{write_gtf}
\alias{write_gtf}
\title{Write a gtf data frames to file}
\usage{
write_gtf(
  gtf_df,
  file,
  header = "##gtf file",
  check_unique = T,
  gzip = F,
  verbose = T
)
}
\arguments{
\item{gtf_df}{gtf data frame}

\item{file}{file path where to save}

\item{header}{vector of header lines}

\item{check_unique}{check for uniqueness of gene_name, gene_id,
transcript_id}
}
\description{
Write a gtf data frames to file
}
