% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xstringset_print.R
\name{xstringset_print}
\alias{xstringset_print}
\title{Print XStringSet to console}
\usage{
xstringset_print(
  set,
  ref = NULL,
  linewidth = 100,
  print_pos = T,
  print_pos_end = F,
  pos_by_ref = T,
  col_out = T,
  call_compare_seq_df_long = F,
  compare_seq_df_long_args = list(seq_original = NULL, change_nonref = T, change_ref = T,
    keep_match_gaps = T)
)
}
\arguments{
\item{set}{XStringSet}

\item{ref}{name of reference sequence}

\item{linewidth}{print width}

\item{print_pos}{print position above}

\item{print_pos_end}{print position of each sequence at lineend (gaps excluded);
requires keep_match_gaps incompare_seq_df_long}

\item{pos_by_ref}{above positions in reference to ref (considering its gaps),
or just total position}

\item{col_out}{colored output?}

\item{call_compare_seq_df_long}{call igsc::compare_seq_df_long to optionally
replace matches and/or mismatches}

\item{compare_seq_df_long_args}{arguments to compare_seq_df_long}
}
\value{
xstringset
}
\description{
XStringSet may contain multiple sequence alignment from DECIPHER::AlignSeqs
}
\examples{
granzymes <- c("GZMA","GZMB","GZMH","GZMK","GZMM")
out <- get_sequences_from_biomart(granzymes)
msaln <- DECIPHER::AlignSeqs(Biostrings::DNAStringSet(setNames(out$coding,
                                                               out$hgnc_symbol)))
# print as is
xstringset_print(msaln)
# print with call_compare_seq_df_long anf GZMB as reference
xstringset_print(msaln, call_compare_seq_df_long = T,
                 ref = "GZMB", print_pos_end = T)
# run compare_seq_df_long  yourself
df <- igsc:::xstringset_to_df(msaln)
msalndf <- compare_seq_df_long(df, ref = "GZMB",
                               seq_original = NULL,
                               change_nonref = T)
msalnx <- df_to_xstringset(msalndf)
# print
xstringset_print(msalnx)
}
